.agent/skills/system-connection-tester/scripts/requirements.txt
psycopg2-binary
redis
paho-mqtt
requests
python-dotenv
